Made unknown module errors well formed
