+++ test/parser/NewlineInString.out
@@ -1,5 +1,5 @@
 Prelude is up to date
 Compiling parser/NewlineInString.hs
-(1,14): Illegal character in string literal
+(1,8): Newline in string literal (expecting ")
   Hint: Correct examples of Strings: "Helium is cool" "abc\ndef" ""
 Compilation failed with 1 error
