+++ test/testOne
@@ -1,7 +1,7 @@
 #!/bin/sh
 echo $1
 /bin/sh compileAndRun $* > result.txt 2>&1
-diff -b result.txt $1.out
+diff --color -b result.txt $1.out
 X=$?
 rm result.txt
 exit $X
